Move trait impls into a `dep_node_key` submodule
